requirements.txt
# Core web framework
fastapi==0.128.0
starlette>=0.21,<1.0
uvicorn==0.23.2

# Pydantic for data validation
pydantic==2.12.5
pydantic_core==2.41.5

# Async HTTP and networking
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
httpcore==1.0.9
httpx==0.28.1
yarl==1.18.0
multidict==6.1.0
frozenlist==1.5.0

# Auth & Security
cryptography==46.0.3
pycparser>=2.23,<3.0
PyJWT==2.10.1
rsa==4.9.1
pywin32-ctypes==0.2.3
python-dotenv==1.2.1

# Google APIs & Firebase
firebase_admin==7.1.0
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
earthengine-api==0.1.389
httplib2==0.31.2
uritemplate==4.2.0

# Discord
discord==2.3.2
discord.py==2.4.0

# Misc utilities
requests==2.32.3
requests-toolbelt==1.0.0
urllib3==2.2.3
certifi==2024.8.30
idna==3.10
charset-normalizer==3.4.0
attrs==24.2.0
anyio==4.12.1
typing-extensions==4.15.0
typing-inspection==0.4.2
typing==3.7.4.3
six==1.16.0
cachecontrol==0.14.4
keyring==25.5.0
SecretStorage==3.5.0
jeepney==0.9.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
readme_renderer==44.0
twine==6.0.1
pyloco==0.0.139
ushlex==0.99.1
matplotlib==3.9.4
numpy==2.1.3
pillow==11.0.0
cycler==0.12.1
kiwisolver==1.4.7
packaging==24.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
fonttools==4.55.3
contourpy==1.3.1
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.2
# propache==0.2.0
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.2
pyasn1-modules==0.4.2
Pygments==2.18.0
matplot==0.1.9
SimpleWebSocketServer==0.1.2
